% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_rza_to_fastrcat.R
\name{join_rza_to_fastrcat}
\alias{join_rza_to_fastrcat}
\title{Link RZA data to FastCAT data}
\usage{
join_rza_to_fastrcat(rza_path, fastrcat_path, depth_range = 5)
}
\arguments{
\item{rza_path}{The path to the directory where the .xlxs rza dataframe
is located along with the .xlxs file name. This is a dataframe specifically
formated for rza data entry.}

\item{fastrcat_path}{The path to the directory of where the .csv file
generated by FastrCAT::make_dataframe_fc(). This must be from the same
Cruise as the RZA data.}

\item{depth_range}{The depth range wanted to define surface and bottom. The
default is 5 meters, this means the surface is defined as 0 to 5
meters and the bottom is defined as maximum tow depth to the 5 meters above.}
}
\value{
writes a new .csv file to the RZA folder of the linked data. Date/time
and latitude/longitude returned will be from the fastrcat.
}
\description{
Links temperature and salinity generated from FastrCAT to
Rapid Zooplankton Assesment data. Six new colums are added to the RZA
data, these are mean of the water column, surface which is defined as the
mean of the top 5 meters, and bottom which is defined as the mean of the
bottom 5 meters; for both temperature and salinity. These are joined
by the station.haul names.
}
