% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_RZA_plots.R
\name{make_rza_plots}
\alias{make_rza_plots}
\title{Abundance Maps for Rapid Zooplankton Assesment taxa.}
\usage{
make_rza_plots(RZA, region, facets = TRUE, Year_folder, Cruise_folder)
}
\arguments{
\item{RZA}{An .xlxs dataframe formated specifically for entry of RZA data.}

\item{region}{There are six geographic regions which produce standardized
maps. The six standardized maps can be set to Arctic, Northern BS, 70 meter,
BASIS, GOA, and MACE 2018.}

\item{facets}{By default facets is set to TRUE. When set to TRUE the
function will produce two sets of faceted maps. One of the 60 bongo rza
taxa and one of the 20 bongo rza taxa. When facets is set to FALSE the
function will produce a map for each rza taxa.}

\item{Year_folder}{The name of the year folder where the rza dataframe
is located.}

\item{Cruise_folder}{The name of the cruise folder where the rza
dataframe is located.}
}
\description{
This function takes an RZA .xlxs dataframe and will make either
individual plots of each RZA taxa group or two faceted maps based on 150
and 500 micron mesh. The maps plot the log abundance per cubic meter of the
rza taxa for each station. The map regions are standardized for each of the
major cruises/grids for EcoFOCI.
}
